+++ js/html-fancy/src/EnvironmentContext.js
@@ -1,0 +1,7 @@
+import {createContext} from 'preact';
+
+
+const EnvironmentContext = createContext({});
+
+export default EnvironmentContext;
+