+++ CHANGES.rst
@@ -6,7 +6,8 @@ envrpt Change Log
 0.2.0 (TBD)
 ===========
 
-* 
+* Added an initial version of the `html-fancy` format, which is a pretty,
+  JavaScript-enhanced HTML report.
 
 
 0.1.0 (2020-02-08)